05/10/2012 Handed in Version.
Takes a filename/s from the commandline and prints data of sin fourier
series (first 4 terms to it).

In commandline type:

>>java HarmonicPlotter giveAnyNametoFile.bla

File can then be opened in graphing program (such as grace) to view
graph.

Instructions for how to do this are given in program

Got 30/30 for this checkpoint.

Got 30/30 for this checkpoint. So might be useful later on for printing
to files.

Will publish checkpoint2 next when get it back.

TODO make number of terms printed an option.
TODO make comments more precise.